requirements.txt
click==6.0
olefile==0.44
Pillow==4.3.0
